fix: 시간표 직접 추가 컴포넌트 구조 변경
