+++ src/pages/Timetable/components/CustomLecture/index.tsx
@@ -1,71 +1,19 @@
-import Listbox from 'components/TimetablePage/Listbox';
 import { ReactComponent as AddIcon } from 'assets/svg/add-icon.svg';
-import { DAYS_STRING, HOUR, MINUTE } from 'static/timetable';
 import styles from './CustomLecture.module.scss';
+import CustomLectureDefaultInput from './CustomLectureDefaultInput';
+import CustomLectureLocationTimeSetting from './CustomLectureLocationTimeSetting';
 
 function CustomLecture() {
   const handleAddLecture = () => {
 
   };
 
-  const onChangeHours = () => {
-
-  };
-
-  const onChangeMin = () => {
-
-  };
   // 해당 파트의 api 명세가 어떻게 나오느냐에 따라 html 구조의 변동이 있을 수 있을 것 같습니다.
   return (
     <form onSubmit={handleAddLecture} className={styles['form-container']}>
-      <div className={styles['form-group']}>
-        <label htmlFor="courseName">
-          <div className={styles['form-group__title']}>
-            수업명
-            <span className={styles['require-mark']}>*</span>
-          </div>
-        </label>
-        <div className={styles['form-group__block']} />
-        <input type="text" id="courseName" name="courseName" placeholder="수업명을 입력하세요." />
-      </div>
-      <div className={styles['form-group']}>
-        <label htmlFor="professorName">교수명</label>
-        <div className={styles['form-group__block']} />
-        <input type="text" name="professorName" placeholder="교수명을 입력하세요." />
-      </div>
-      <div className={styles['form-group-time']}>
-        <label htmlFor="place">
-          <div className={styles['form-group__title']}>
-            시간
-            <span className={styles['require-mark']}>*</span>
-          </div>
-        </label>
-        <div className={styles['form-group-time__container']}>
-          <div className={styles['form-group-time__weekdays']}>
-            {DAYS_STRING.map((weekday) => (
-              <div key={weekday}>
-                <button type="button" className={styles['form-group-time__weekdays-button']}>{weekday}</button>
-              </div>
-            ))}
-          </div>
-          <div className={styles['form-group-time__time']}>
-            <div className={styles['form-group-time__time-section']}>
-              <Listbox list={HOUR} value="00시" onChange={onChangeHours} version="addLecture" />
-              <Listbox list={MINUTE} value="00분" onChange={onChangeMin} version="addLecture" />
-            </div>
-            <span>-</span>
-            <div className={styles['form-group-time__time-section']}>
-              <Listbox list={HOUR} value="00시" onChange={onChangeHours} version="addLecture" />
-              <Listbox list={MINUTE} value="00분" onChange={onChangeMin} version="addLecture" />
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className={styles['form-group']}>
-        <label htmlFor="place">장소</label>
-        <div className={styles['form-group__block']} />
-        <input type="text" name="place" placeholder="장소를 입력하세요." />
-      </div>
+      <CustomLectureDefaultInput title="수업명" placeholder="수업명을 입력하세요." require />
+      <CustomLectureDefaultInput title="교수명" placeholder="교수명을 입력하세요." require={false} />
+      <CustomLectureLocationTimeSetting />
       <button type="button" className={styles['form-group-add-button']}>
         <span>시간 및 장소 추가</span>
         <AddIcon />
